--- a/docs/changes/coordinates/16135.feature.rst
+++ b/docs/changes/coordinates/16135.feature.rst
@@ -1,0 +1,2 @@
+Added a fast path for converting ``PhysicsSphericalRepresentation`` to
+``CylindricalRepresentation``.