@nstarman 2024-03-01
Including the reverse makes sense to me. I just encountered this in my research and noted it was going through Cartesian, which seemed unnecessary since only cos and sin theta need to be computed to actually do the transform.