@mhvk 2024-03-01
> Why coverage bot says your new code isn't covered?

It might actually be that we do not explicitly test PhysicsSpherical to Cylindrical...

> p.s. Oooo I wonder if #15779 would be useful here.

Yes, I'm looking forward very much to have benchmarks again!