@neutrinoceros 2024-03-04
@nstarman this is due to pytest 8.1, see #16146