@pllim 2024-03-04
Re: CI -- pytest yanked 8.1. I restarted the jobs and unless you broke it, they should pass now (except devdeps).