@pllim 2024-03-16
I just applied a label so this runs benchmark implemented in https://github.com/astropy/astropy/pull/15779 (cc @MridulS) assuming this functionality is already benchmarked over at https://github.com/astropy/astropy-benchmarks/tree/main/benchmarks ?