@pllim 2024-03-16
Does the benchmark log help at all?

https://github.com/astropy/astropy/actions/runs/8304528883/job/22730265639?pr=16135