@pllim 2024-03-16
@MridulS , the log actually says

```
SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
Error: PERFORMANCE DECREASED.
```

but the job is green. Is this expected?