@MridulS 2024-03-16
Ahhhh. No it's supposed to give a red cross. I guess in the new iterations I removed that logic. I'll send a PR tomorrow to fix that.