@mhvk 2024-03-18
Was looking back at this PR to see its status: @eerovaher - are you happy with @nstarman's changes following your comments?

On the benchmarks, in principle it would be bad if this *decreased* performance, but looking at the log, I see
```
| Change   | Before [d3a39fa1]    | After [ce2f5fbb]    |   Ratio | Benchmark (Parameter)          |
|----------|----------------------|---------------------|---------|--------------------------------|
| +        | 54.7±10μs            | 209±8μs             |    3.81 | units.time_quantity_times_unit |
```
This is indeed a large change but has absolutely nothing to do with this PR. Which in itself is perhaps worrying? (Perhaps best discussed elsewhere, though!)