@pllim 2024-03-18
Hmm. I don't see a big jump here:

https://spacetelescope.github.io/bench/astropy-benchmarks/#units.time_quantity_times_unit