@mhvk 2024-03-18
Sounds like it was a fluke. Since the test was in `units`, unrelated to the coordinates stuff here, I think there is no need to re-test. Though we probably need some representation tests, so one can see the benefit of this PR!