@eerovaher 2024-03-19
I'm not seeing any performance increases in the benchmark suite. The most likely explanation is that the `PhysicsSphericalRepresentation` to `CylindricalRepresentation` conversion is not included in the benchmarks, but I don't know what is actually included and because the pull request that implemented benchmarking was merged without any documentation then I can't look it up either.