@pllim 2024-03-19
The PR benchmark basically runs https://github.com/astropy/astropy-benchmarks/tree/main/benchmarks but in a *relative* way instead of absolute.

Not sure how much this issue helps to explain but it has links: https://github.com/astropy/astropy-benchmarks/issues/106