@nstarman 2024-03-21
On my computer I consistently get a 5% speed boost.