@nstarman 2024-03-21
Also, I discovered there's another good reason for this PR.

Pre this PR
```python
import astropy.units as u
import astropy.coordinates as cs

>>> cyl = cs.CylindricalRepresentation(rho=0 * u.kpc, phi=270 * u.deg, z=3 * u.kpc)
>>> cyl.represent_as(cs.PhysicsSphericalRepresentation).phi.deg
180.0
```

With this PR
```python
import astropy.units as u
import astropy.coordinates as cs

>>> cyl = cs.CylindricalRepresentation(rho=0 * u.kpc, phi=270 * u.deg, z=3 * u.kpc)
>>> cyl.represent_as(cs.PhysicsSphericalRepresentation).phi.deg
270.0
```

Consider the transformation (https://en.wikipedia.org/wiki/Cylindrical_coordinate_system)

<img src="https://github.com/astropy/astropy/assets/8949649/980b922a-4492-44da-bf6a-261e466614b4" width="300" />

The `phi` value is conserved through the transformation. Technically at `r=0`, `phi` is not well defined, however I think that keeping `phi` is better than changing it to a value determined by the `Cartesian -> Spherical` conventions.