@eerovaher 2024-03-21
I agree with that reasoning. Do add that example to the tests.