@mhvk 2024-03-21
On the performance: I think there is a reasonable benefit only when no differentials are involved - those needs a lot more calculations and are not improved by this PR.